feat: add stage backend publish task
