+++ src/dashboard/apigateway/apigateway/biz/backend.py
@@ -19,7 +19,9 @@ from typing import Any, Dict
 
 from django.db import transaction
 
-from apigateway.core.models import Backend, BackendConfig
+from apigateway.controller.tasks.syncing import trigger_gateway_publish
+from apigateway.core.constants import DEFAULT_BACKEND_NAME, PublishSourceEnum
+from apigateway.core.models import Backend, BackendConfig, Proxy
 
 
 class BackendHandler:
@@ -58,7 +60,8 @@ class BackendHandler:
     def update(backend: Backend, data: Dict[str, Any], updated_by: str) -> Backend:
         """更新后端服务"""
         backend.type = data["type"]
-        backend.name = data["name"]
+        if backend.name != DEFAULT_BACKEND_NAME:
+            backend.name = data["name"]
         backend.description = data["description"]
         backend.updated_by = updated_by
         backend.save()
@@ -69,11 +72,33 @@ class BackendHandler:
         backend_configs = []
         for config in data["configs"]:
             backend_config = stage_configs[config["stage_id"]]
-            backend_config.config = {key: value for key, value in config.items() if key != "stage_id"}
+            new_config = {key: value for key, value in config.items() if key != "stage_id"}
+            if new_config == backend_config.config:
+                continue
+
+            backend_config.config = new_config
             backend_config.updated_by = updated_by
 
             backend_configs.append(backend_config)
 
         BackendConfig.objects.bulk_update(backend_configs, fields=["config", "updated_by"])
 
+        # 触发变更的stage的发布流程
+        for backend_config in backend_configs:
+            trigger_gateway_publish(
+                PublishSourceEnum.BACKEND_UPDATE,
+                updated_by,
+                backend_config.gateway_id,
+                backend_config.stage_id,
+                is_sync=True,
+            )
+
         return backend
+
+    @staticmethod
+    def deletable(backend: Backend) -> bool:
+        """判断后端服务是否可删除"""
+        if backend.name == DEFAULT_BACKEND_NAME:
+            return False
+
+        return not Proxy.objects.filter(backend=backend).exists()
