+++ src/dashboard/apigateway/apigateway/biz/stage.py
@@ -21,7 +21,8 @@ from typing import Any, Dict, Optional
 
 from django.db import transaction
 
-from apigateway.core.constants import DEFAULT_BACKEND_NAME, DEFAULT_STAGE_NAME, StageStatusEnum
+from apigateway.controller.tasks.syncing import trigger_gateway_publish
+from apigateway.core.constants import DEFAULT_BACKEND_NAME, DEFAULT_STAGE_NAME, PublishSourceEnum, StageStatusEnum
 from apigateway.core.models import Backend, BackendConfig, MicroGateway, Release, ReleaseHistory, Stage
 from apigateway.utils.time import now_datetime
 
@@ -102,7 +103,11 @@ class StageHandler:
         stage.updated_by = updated_by
         stage.save()
 
-        # TODO 下架/上架发布
+        if status == StageStatusEnum.INACTIVE.value:
+            # 触发环境发布
+            trigger_gateway_publish(
+                PublishSourceEnum.STAGE_DISABLE, updated_by, stage.gateway_id, stage.id, is_sync=True
+            )
 
     @staticmethod
     def delete_by_gateway_id(gateway_id):
